Add flag --input_type. Add binary input. Add decoder from bin to str.
git-svn-id: http://data-race-test.googlecode.com/svn/trunk@2382 93a3164d-3a44-0410-8ae5-4baa038b0729

+++ tsan/ts_util.h
@@ -46,9 +46,9 @@ extern void ThreadSanitizerDumpAllStacks();
    ThreadSanitizerDumpAllStacks(); \
    exit(1); }} while (0)
 #elif defined(TS_OFFLINE)
-extern int offline_line_n;
+extern unsigned long offline_line_n;
 # define CHECK(x) do { if (!(x)) { \
-    Printf("ASSERT on line %d\n", offline_line_n); \
+    Printf("ASSERT on line %ld\n", offline_line_n); \
      assert(x);}} while (0)
 #else
 # define CHECK assert
